refactor(server): improved naming
