+++ src/server/routes/lumi-h5p.ts
@@ -2,8 +2,6 @@ import express from 'express';
 
 import lumiController from '../controller/lumi-h5p';
 import Logger from '../helper/Logger';
-import h5p from '../h5p';
-import { IUser } from 'h5p-nodejs-library';
 
 const log = new Logger('routes:lumi-h5p');
